Refactor avatar handling.
PhpBB's driver->get_data() method does not expose absolute URLs and
we need special handling for every avatar type due to each using a different URL pattern.
We therefore have to hardcode the avatar type name and path fragments and depend on
internals with which we will need to maintain compatibility in case of future upstream changes.
Using our own implementation with less overhead does not introduce further costs in this case.